python-api/requirements.txt
cherrypy
cherrypy-cors
numpy
pydicom
matplotlib
six
Pillow
#Shapely
scikit-image
jsonmerge
deepmerge
requests
nibabel
dicom
pyelastix
scipy
dicom-contour 
python-dotenv
pyelastix
scikit-image

imageio
ipympl
dicompyler-core